requirements.txt
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
dnspython==2.6.1
email_validator==2.2.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.0
idna==3.8
importlib-metadata==6.11.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.7.0
numpy==1.26.4
packaging==23.2
pandas==2.2.2
Pillow==9.5.0
protobuf==4.25.4
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
pywin32==306
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
setuptools==74.1.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.34
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.3
validators==0.34.0
watchdog==4.0.2
zipp==3.20.2
zope.interface==7.0.3